+++ service/src/main/java/com/itclopedia/courses/services/ProjectService.java
@@ -1,7 +1,13 @@
 package com.itclopedia.courses.services;
 
 import com.itclopedia.courses.dao.ProjectRepository;
+import com.itclopedia.courses.dao.UserRepository;
+import com.itclopedia.courses.dto.ProjectDTO;
+import com.itclopedia.courses.exceptions.EntityAlreadyExistsException;
+import com.itclopedia.courses.exceptions.EntityNotFoundException;
+import com.itclopedia.courses.mapper.ProjectMapper;
 import com.itclopedia.courses.models.Project;
+import com.itclopedia.courses.models.User;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,25 +17,40 @@ import org.springframework.stereotype.Service;
 public class ProjectService {
 
     private final ProjectRepository projectRepository;
-
+    private final UserRepository userRepository;
+    private final ProjectMapper projectMapper = ProjectMapper.INSTANCE;
     @Autowired
-    public ProjectService(ProjectRepository projectRepository) {
+    public ProjectService(ProjectRepository projectRepository, UserRepository userRepository) {
         this.projectRepository = projectRepository;
+        this.userRepository = userRepository;
     }
 
-    public void addProject(Project project) {
+    public void addProject(ProjectDTO projectDTO) {
+        Project project = projectMapper.toProject(projectDTO, userRepository);
+        User user = userRepository.findById(projectDTO.getUserId())
+                .orElseThrow(() -> new EntityAlreadyExistsException("User", projectDTO.getUserId()));
+        project.setUser(user);
         projectRepository.save(project);
     }
 
-    public void updateProject(Project project) {
+    public void updateProject(ProjectDTO projectDTO) {
+        Project project = projectMapper.toProject(projectDTO, userRepository);
+        User user = userRepository.findById(projectDTO.getUserId())
+                .orElseThrow(() -> new EntityNotFoundException("User", projectDTO.getUserId()));
+        project.setUser(user);
         projectRepository.save(project);
     }
 
-    public Project getProjectById(int id) {
-        return projectRepository.findById(id).orElse(null);
+    public ProjectDTO getProjectById(int id) {
+        Project project = projectRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Project", id));
+        return projectMapper.toProjectDTO(project);
     }
 
     public void deleteProject(int id) {
+        if (!projectRepository.existsById(id)) {
+            throw new EntityNotFoundException("Project", id);
+        }
         projectRepository.deleteById(id);
     }
 }
